fix(mir): fail closed on Vec elements the layout registries cannot see
The Vec-element classifier's `Unsupported(UnenumeratedShape)` sentinel was
mapped back to a buffer-only release at both consumers (the definition-site
destructor recipe and the release-symbol verdict), so an element the
heap-ownership authority never saw was freed as if proven heap-free. The
authority answers "no obligation" for an unregistered `Named` head exactly as
for a heap-free one, which made that arm a fail-open.

The classifier now decides the whole domain itself: an element whose every
head is visible to a registry and carries no obligation is `Plain` (a
heap-free generic record instance, a bare runtime handle, a generic
skeleton's free type parameter — the shapes the sentinel actually covered,
measured over the corpus, the vertical-slice fixtures, and the dogfood
modules); an element with an unseen `Named` head is
`Unsupported(UnknownValueClass)` and publishes no recipe, gets no release
symbol, and is rejected at compile. The sentinel and both fallback arms are
deleted, and `binding_ty_is_plain_vec` is a projection of the classifier
again.

The element classification moves to `expr/vec_element_release.rs` to keep
`expr.rs` under its line ceiling.

Function-entry cancellation keeps the `ParamBoundary` fact as its authority;
the measurement showing the event stream cannot name that set (guarded owned
parameters versus a guarded cursor mint over a borrowed `var self`) is
recorded beside the function and in the ladder.
