+++ LESSONS.md
@@ -615,3 +615,15 @@ Nuance row (table under the core):
 - **two traps, both hit**: (1) "ended" must exclude `Join` inputs and `Reset`/`Rearm` predecessors — those are renames, and dropping a `Join` input on its edge destroys the joined `var` (goobers `workflow_machine` went to 48 "join does not enumerate its exact possible incoming owners" findings before this was excluded). (2) The edge drop must be gated on local liveness at the join: `durable.parent` returns a `var` string through an if-expression result slot with a `Move` to `ReturnSlot` and no `Transfer`, so without the gate the owning edge freed the value about to be returned.
 - **the residue is a producer hole, not a checker gap**: a verifier arm reporting "consumed on some paths and still owned on others" was written and then withdrawn — it fires on three real pre-existing lowering holes (return-tail `Mint` where a `Join`/`Transfer` belongs, in `try_encode`-shaped returns and `durable.parent`; and the guarded match arm that mints its binder before the guard decides, #3114). Turning it on without those producer fixes turns documented leaks into hard errors on std (`fs.try_read`, `process.last_process_error`, `stream.from_file`) and goobers. Land the producer fixes first, then the rule.
 - **evidence**: `tests/vertical-slice/accept/vec_iter_yield_owned_conditional_move.hew` (formerly a reject fixture behind the deleted NYI wall) prints `n=3` twice and `conditional-move-ok`; `vec_iter_clone_out::vec_iter_yield_conditionally_consumed_is_released_on_its_owning_edge`; `conditional_consume_release::tests` (inline Goto release, split call edge, read-after-join emits nothing, redefinition admitted, `Join` input untouched, guarded untouched).
+
+### `vec-element-unknown-head-fails-closed` — "no drop obligation" from the heap authority is not evidence for a head it cannot see
+
+- **area**: `Builder::classify_vec_element_release` (`hew-mir/src/lower/expr.rs`), `owner_definition_drop_recipe`, `vec_release_symbol_verdict`
+- **invariant**: `ty_carries_drop_obligation` answers `NONE` for an unregistered `Named` head exactly as it does for a heap-free one. A `Vec` element classification may therefore call the element `Plain` only when every head in it is visible to a registry (`ty_has_unregistered_named` is false); an unseen head is `Unsupported(UnknownValueClass)` — no `DropRecipe`, `Unwired` release symbol, rejected by `unsupported_vec_element_in_ty`. No consumer maps an `Unsupported` classification back to a release symbol (the old `UnenumeratedShape => VecPlain` arms were a fail-open the classifier's own doc called fail-closed).
+- **what the sentinel actually covered** (probe over corpus + fixtures + goobers, F1 round 5): free `TypeParam`s in generic origin skeletons, `std.arena.Slot<T>`/`Slot<i64>`/`Slot<Point>` (registered heap-free generic records the value-class table never marks `BitCopy`), and `LocalPid<Actor>` (a builtin handle). All are heap-free by a registry that saw them, so they classify `Plain` now with unchanged runtime behaviour; a `Vec<T>` skeleton keeps its outer-buffer recipe (`generic_vec_definition_publishes_outer_buffer_recipe`).
+- **cousin, not a duplicate**: `binding_ty_is_plain_vec` is a projection of the classifier again; `nested_tuple_with_unregistered_named_is_not_admitted` holds because the unregistered head is now `Unsupported`, not because the plain gate is a separate copy.
+
+### `entry-cancel-needs-the-boundary-mode` — the event stream cannot name the function-entry cancellation set
+
+- **area**: `entry_cancel_parameter_owners` (`hew-mir/src/lower/drop_plan.rs`)
+- **finding**: two replay-only candidates were measured against the `ParamBoundary`-mode rule across ~28,700 functions. "Every `Mint` over a parameter local in the entry block" adds the guarded cursor `Mint` that `VecIter::next$$…`'s `var self` publishes over a `BorrowReadOnly` parameter (3 functions; the caller owns that cursor, dropping it at entry is a double free). "Every UNGUARDED such `Mint`" loses ~400 owned parameters whose `Mint` is guarded, which must be dropped unguarded on this edge because codegen injects the cancellation before the guard sidecar is initialised. The boundary mode is the only fact that separates the two, and it is part of Checked MIR's decision stream, so the exception stays and is documented in §2.7.
