+++ hew-mir/src/lower/drop_plan/drop_admission_type_shape_pins.rs
@@ -446,12 +446,9 @@ fn collection_handle_predicate_projects_from_heap_leaf() {
 /// (`Unwired`): the per-element release for those shapes is unwired, so
 /// every consulting site must refuse the construct at compile time —
 /// never emit the buffer-only `hew_vec_free` over owned element nodes.
-/// The residual `Unsupported(UnenumeratedShape)` sub-domain deliberately
-/// keeps the buffer-only verdict, drawing the same boundary as the compile
-/// reject `unsupported_vec_element_walk`:
-///   - `UnenumeratedShape` (`Vec<T>` unsubstituted): the element owns no
-///     heap as a flat element, so the buffer free IS the complete
-///     release — refusing would reject un-monomorphised generic bodies;
+/// An un-monomorphised `Vec<T>` classifies `Plain` (the skeleton's outer
+/// buffer release), and a `Named` element no registry can see is
+/// `Unsupported(UnknownValueClass)` — `Unwired`, never a buffer-only symbol.
 ///
 /// A registered heap-owning record observed without this function's
 /// harvest key is instead classified harvest-independently and released
@@ -542,16 +539,20 @@ fn yield_and_field_pickers_match_legacy_symbol_table() {
             Unwired(FailClosedReason::NoReleaseProtocol),
             Unwired(FailClosedReason::NoReleaseProtocol),
         ),
-        // Vec arm — the residual Unsupported sub-domain that keeps the
-        // buffer-only verdict (the boundary
-        // `unsupported_vec_element_walk` draws; see the test doc).
+        // Vec arm — a generic skeleton's `Vec<T>` releases its outer buffer.
         (
-            "Vec<T> unsubstituted (Unsupported/UnenumeratedShape)",
+            "Vec<T> unsubstituted (Plain)",
             vec_of(ResolvedTy::TypeParam {
                 name: "T".to_string(),
             }),
             Wired("hew_vec_free"),
             Wired("hew_vec_free"),
+        ),
+        (
+            "Vec<Ghost> unregistered head (Unsupported/UnknownValueClass)",
+            vec_of(named("Ghost")),
+            Unwired(FailClosedReason::UnknownValueClass),
+            Unwired(FailClosedReason::UnknownValueClass),
         ),
         (
             "Vec<HeapRow> unharvested (Unsupported/NoReleaseProtocol, owned-ABI releasable)",
@@ -619,7 +620,7 @@ fn yield_and_field_pickers_match_legacy_symbol_table() {
     }
 
     // The Unsupported rows above carry exactly the two fail-closed
-    // reasons: the unwired release protocols and the anti-drift sentinel.
+    // reasons: the unwired release protocols and the unseen head.
     assert_eq!(
         builder.classify_vec_element_release(&ResolvedTy::Bytes),
         VecElementRelease::Unsupported(FailClosedReason::NoReleaseProtocol)
@@ -632,7 +633,11 @@ fn yield_and_field_pickers_match_legacy_symbol_table() {
         builder.classify_vec_element_release(&ResolvedTy::TypeParam {
             name: "T".to_string(),
         }),
-        VecElementRelease::Unsupported(FailClosedReason::UnenumeratedShape)
+        VecElementRelease::Plain
+    );
+    assert_eq!(
+        builder.classify_vec_element_release(&named("Ghost")),
+        VecElementRelease::Unsupported(FailClosedReason::UnknownValueClass)
     );
     // The releasable-boundary row rides `NoReleaseProtocol` too — it is
     // the `elem_is_owned_abi_releasable` exclusion, not the reason, that
